Implement assert_enqueued_async_subscriber
