+++ test/active_event_store/test_helper_test.rb
@@ -4,15 +4,23 @@ require "test_helper"
 require "active_event_store/test_helper"
 require_relative "../../spec/support/test_events"
 
-class TestHelperTest < Minitest::Test
+class ActiveEventStore::TestHelperTest < Minitest::Test
   include ActiveEventStore::TestHelper
   attr_reader :event_class,
     :event,
     :event2
 
+  module TestSubscriber
+    class << self
+      def call(_event)
+      end
+    end
+  end
+
   def setup
     super
     @event_class = ActiveEventStore::TestEvent
+    @test_subscriber = TestSubscriber
     @event = event_class.new(user_id: 25, action_type: "birth")
     @event2 = event_class.new(user_id: 1, action_type: "death")
   end
@@ -96,5 +104,24 @@ class TestHelperTest < Minitest::Test
   def test_refute_event_published_on_no_events
     refute_event_published(ActiveEventStore::TestEvent, with: {user_id: 25, action_type: "death"}) do
     end
+  end
+
+  # == async_subscriber tests
+  def test_assert_async_event_subscriber_with_event
+    ActiveEventStore.subscribe(@test_subscriber, to: event_class, sync: false)
+
+    assert_async_event_subscriber_enqueued(@test_subscriber) do
+      ActiveEventStore.publish(event)
+    end
+  end
+
+  def test_assert_async_event_subscriber_raises_without_event
+    ActiveEventStore.subscribe(@test_subscriber, to: event_class, sync: false)
+
+    e = assert_raises Minitest::Assertion do
+      assert_async_event_subscriber_enqueued(@test_subscriber) {}
+    end
+
+    assert_match /No enqueued job found.*job.*ActiveEventStore::TestHelperTest::TestSubscriber::SubscriberJob.*queue.*"events_subscribers"/, e.message
   end
 end
